audio_player/requirements.txt
pygame~=2.0.1
python-vlc~=3.0.12118
